requirements.txt
feedparser==6.0.11
python-telegram-bot==21.4
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
